# Optimize getting the latest events by type per room